New folder (28)/requirements.txt
python-telegram-bot==13.15
websocket-client>=1.6.0
pandas>=1.5.0
ta>=0.10.0
python-dotenv>=1.0.0
pytz>=2023.3
